LPS-108239 Move dbPartition classes to the proper package
