+++ portal-impl/src/com/liferay/portal/dao/init/DBInitUtil.java
@@ -17,7 +17,7 @@ package com.liferay.portal.dao.init;
 import com.liferay.petra.io.StreamUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.events.StartupHelperUtil;
-import com.liferay.portal.internal.db.partition.DBPartitionHelperUtil;
+import com.liferay.portal.db.partition.DBPartitionHelperUtil;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManager;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
